ai-service/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
scikit-learn==1.3.0
pandas==2.1.0
numpy==1.25.2
python-dotenv==1.0.0
pydantic==2.3.0
